requirements.txt
RPi.GPIO==0.7.1
hx711py==1.0.1
openai==1.3.0
flask==2.3.3
flask-cors==4.0.0 